+++ src/main/java/ai/mcpdirect/studio/tool/openapi/OpenAPIToolProvider.java
@@ -97,7 +97,7 @@ public class OpenAPIToolProvider extends OpenAPIServer implements AIToolProvider
     public void refreshTools() {
     }
     public void close(){
-        status=STATUS_ABANDONED;
+        status=STATUS_OFF;
     }